remove unwanted files
